feat: Add a servings selectBox to the app
Add servings selectBox to the app and modified the calories data to display calories data as per the formula calories * number of servings.
Modify the name of the id property in model to item_id as it is is being confused with the original id for each data. item_id was not adding the same item twice.